backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
pyjwt
prometheus-flask-exporter
werkzeug